+++ FIFA_API/Controllers/Base/CommandesController.cs
@@ -9,6 +9,7 @@ using FIFA_API.Models.EntityFramework;
 using FIFA_API.Utils;
 using FIFA_API.Models;
 using Microsoft.AspNetCore.Authorization;
+using FIFA_API.Models.Controllers;
 
 namespace FIFA_API.Controllers
 {
@@ -38,7 +39,7 @@ namespace FIFA_API.Controllers
         // GET: api/Commandes/5
         [HttpGet("{id}")]
         [Authorize(Policy = Policies.User)]
-        public async Task<ActionResult<Commande>> GetCommande(int id)
+        public async Task<ActionResult<CommandeDetails>> GetCommande(int id)
         {
             Utilisateur? user = await this.UtilisateurAsync();
             if (user is null) return Unauthorized();
@@ -46,9 +47,9 @@ namespace FIFA_API.Controllers
             var commande = await _context.Commandes.GetByIdAsync(id);
 
             if (commande is null) return NotFound();
-            if(!User.IsInRole(MANAGER_POLICY) && commande.Id != user.Id) return NotFound(); 
+            if(!User.IsInRole(MANAGER_POLICY) && commande.Id != user.Id) return NotFound();
 
-            return commande;
+            return await CommandeDetails.FromCommande(commande, _context);
         }
 
         // PUT: api/Commandes/5
