+++ FIFA_API/Models/EntityFramework/StockProduit.cs
@@ -1,8 +1,9 @@
 using FIFA_API.Models.Annotations;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using System.Text.Json.Serialization;
 
-﻿namespace FIFA_API.Models.EntityFramework
+namespace FIFA_API.Models.EntityFramework
 {
 	[Table("t_j_stockproduit_spr")]
     [ComposedKey(nameof(IdVCProduit), nameof(IdTaille))]
@@ -14,10 +15,10 @@ using System.ComponentModel.DataAnnotations.Schema;
         [Column("tpr_id", Order = 1), Required]
         public int IdTaille { get; set; }
 
-        [ForeignKey(nameof(IdVCProduit))]
+        [ForeignKey(nameof(IdVCProduit)), JsonIgnore]
         public VarianteCouleurProduit VCProduit { get; set; }
 
-        [ForeignKey(nameof(IdTaille))]
+        [ForeignKey(nameof(IdTaille)), JsonIgnore]
         public TailleProduit Taille { get; set; }
 
 		[Column("spr_stocks")]
